python-ai-service/requirements.txt
Flask
Flask-CORS
opencv-python
tensorflow
Pillow
numpy
scikit-learn
matplotlib